% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGO.R
\name{annotateGO}
\alias{annotateGO}
\title{Map genes to annotated GO terms}
\usage{
annotateGO(k)
}
\arguments{
\item{k}{Vector of gene to map to GOIDs}
}
\value{
Dataframe with GOIDs, terms and genes
}
\description{
Map genes to annotated GO terms
}
\examples{
annotateGO(c("MAPK1", "MAPK3"))
}
