% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOdistr.R
\name{GOdistr}
\alias{GOdistr}
\title{Count number of times a term is annotated within gene sample.}
\usage{
GOdistr(genes, annots = annotations)
}
\arguments{
\item{genes}{vector of hgnc gene names}

\item{annots}{output of annotateGO()}
}
\value{
dataframe of GO term counts
}
\description{
To count the number of times a term occurs in a given sample

Precursor of hypergeometric distribution calculation.
}
