% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrichment.R
\name{GOenrichment}
\alias{GOenrichment}
\title{Over Representation Analysis}
\usage{
GOenrichment(sampleDistr, sampleSize, bkgdDistr, bkgdSize, p = 0.05)
}
\arguments{
\item{sampleDistr}{Output of GOdistr() for sample}

\item{sampleSize}{Size of sample}

\item{bkgdDistr}{Output of GOdistr() for background population}

\item{bkgdSize}{Size of background population}

\item{p}{Significance threshold}
}
\value{
Over represented GO terms as df
}
\description{
Calculate over enrichment of GO terms in

a sample Explanation can be found in:

https://mengnote.blogspot.co.uk/2012/12/calculate-correct-hypergeometric-p.html
}
