% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSTRINGnw.R
\name{importSTRINGnw}
\alias{importSTRINGnw}
\title{Title Create network from STRING databases}
\usage{
importSTRINGnw(conf, speciesID = 9606)
}
\arguments{
\item{conf}{Confidence threshold for STRING edges}

\item{speciesID}{NCBI species ID. Default Homo sapiens, 9606}
}
\value{
Edgelist of human STRING network
}
\description{
Title Create network from STRING databases
}
\examples{
importSTRINGnw(700)
}
