% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuzz-ggplot.R
\name{mFuzz.ggplot}
\alias{mFuzz.ggplot}
\title{Fuzy C-Means Clustering and Visualisation}
\usage{
mFuzz.ggplot(data, clustering, centre = F, sort.columns = T)
}
\arguments{
\item{data}{a dataframe or matrix}

\item{clustering}{mFuzz output}

\item{centre}{logical; should cluster centers be plotted}

\item{sort.columns}{logical; should column names be replaced by numbered numbers}
}
\value{
list containing ggplot object and dataframe of clustering values
}
\description{
From Jamie Soul; adapted from jspaezp on GitHub
}
\examples{
data <- matrix(rnorm(1:300), nrow = 100, ncol = 3)
}
