% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyGOReqData.R
\name{simplifyGOReqData}
\alias{simplifyGOReqData}
\title{Gather data required for simplify() function, to reduce
semantically similar GO terms}
\usage{
simplifyGOReqData()
}
\value{
List of dataframes
}
\description{
Gather data required for simplify() function, to reduce
semantically similar GO terms
}
\examples{
simplifyGOReqData()
}
