% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRINGvis.R
\name{STRINGvis}
\alias{STRINGvis}
\title{Construct a ggraph visualisation and igraph object
from HGNC names and STRING network.}
\usage{
STRINGvis(genes, stringNW)
}
\arguments{
\item{genes}{a vector of HGNC identifiers}

\item{stringNW}{Output of importSTRINGnw}
}
\value{
list of ggraph visualisation and igraph network
}
\description{
Construct a ggraph visualisation and igraph object
from HGNC names and STRING network.
}
