% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOenrichment.R
\name{GOenrichment}
\alias{GOenrichment}
\title{Over Representation Analysis}
\usage{
GOenrichment(
  GO.clust,
  sample,
  HeLa.bkgd = 12427,
  HeLa.GO = GO.distr.HeLa,
  p = 0.05
)
}
\arguments{
\item{GO.clust}{}

\item{sample}{}

\item{HeLa.bkgd}{}

\item{HeLa.GO}{}

\item{p}{}
}
\value{

}
\description{
Calculate over enrichment of GO terms in

a sample Explanation can be found in:

https://mengnote.blogspot.co.uk/2012/12/calculate-correct-hypergeometric-p.html
}
