% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeTruncNorm.R
\name{imputeTruncNorm}
\alias{imputeTruncNorm}
\title{Title Impute Data by drawing randomly from a distribution}
\usage{
imputeTruncNorm(v)
}
\arguments{
\item{v}{a vector containing NA}
}
\value{
imputed values
}
\description{
The default represents a stimulation to represent a population

below the limits of detection, as described in Tyanova et al., 2016
}
\examples{
foo <- rnorm(100)
foo[sample(100, 5)] <- NA
imputeTruncNorm(foo)

}
